requirements.txt
fastapi
uvicorn
openai
langchain
langchain-openai
pydantic
python-dotenv
faiss-cpu
llama-index-core
llama-index-vector-stores-faiss
streamlit
